rename modification requested notification usecase and add migration
